requirements.txt
asgiref==3.6.0
autopep8==2.0.1
boto3==1.26.69
botocore==1.29.69
colorama==0.4.6
Django==4.1.6
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.1.1
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
whitenoise
